pyproject.toml
[project]
name = "ai-research-assistant"
version = "0.1.0"
description = "AI Research Assistant for Technical PDFs using RAG and MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "mcp>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
research-assistant = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
